requirements.txt
aiogram==3.16.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
aiosqlite==0.20.0
aiohttp==3.11.11
openai==1.58.1
